refactored codebase to improve clarity
