+++ .env.example
@@ -2,6 +2,7 @@ GOOGLE_GENAI_USE_VERTEXAI=1
 GOOGLE_CLOUD_PROJECT=portfolio-builder-494211
 GOOGLE_CLOUD_LOCATION=us-central1
 STITCH_API_KEY=your-stitch-api-key
+CV_GOOGLE_DOC_ID=your-google-doc-file-id
 
 # Gmail SMTP (used by both demo and production pipelines)
 GMAIL_USER=your-gmail@gmail.com
